Extract Proto ENUM
